refactor(gui): extract diff sequence info to new class
Simplifying the FrameNavigation class further.
