+++ GUI/src/main/kotlin/frameNavigation/FrameNavigationInterface.kt
@@ -25,9 +25,4 @@ interface FrameNavigationInterface {
      * Jump to the next diff
      */
     fun jumpToNextDiff(forward: Boolean)
-
-    /**
-     * Get count of frames in diff
-     */
-    fun getSizeOfDiff(): Int
 }
